Move heuristic evaluation to evaluation folder.
